requirements.txt
# Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database - MSSQL
mssql-django==1.3
pyodbc==5.0.1

# Environment Variables
python-dotenv==1.0.0

# Google Vision API (OCR)
google-cloud-vision==3.4.5

# OpenAI (ChatGPT)
openai==1.3.5

# Image Processing
Pillow==10.1.0

# Utilities
python-dateutil==2.8.2
